// src/app/models/sjcs/FundamentoResolucionItem.ts
export class FundamentoResolucionItem {
  idFundamento: string;
  descripcionFundamento: string;
  idInstitucion: string;
  fechaModificacion: string;
  idTipoResolucion: string;
  descripcionResolucion: string;
  textoPlantilla: string;
  textoPlantilla2: string;
  textoPlantilla3: string;
  textoPlantilla4: string;
  historico: boolean;
  fechaBaja: string;
  codigoExt: string;
  codigoEjis: string;
  constructor() { }
}